feat(performance): Materialize metadata fields
To improve performance when defining views on metadata fields, they will
now be materialized in the metadata_fields table. A task is scheduled on
@reboot to perform the initial refresh of this table. Subsequently, the
table will be refreshed during bulk load of structure metadata. The
latest updated_at in the metadata_fields table will be used as a "low
watermark" to scan only new metadata in data_structure_versions and
structure_metadata.

JIRA: TD-4447