refactor: Implementation structure indexing
Previous to this change (and since TD-5001), any aliased data structure
referenced by an implementation was indexed with it's alias in the
'name' property. This caused an issue when executing quality
implementations, as the technical name is required, not the alias
(functional name).

This change indexes the data structure's original name in the 'name'
property, and it's alias (if present) in the new 'alias' property of the
embedded document.

JIRA: TD-5344